* started email feature

* email function

* email function

* re-converted option #1 into a checkbox

* Convert php_notifier_send_email option to boolean

* Return early, prevent email error when up to date version of PHP. Resolves #6. Update readme.md

* Add grunt menu, update grunt tasks, test grunt tasks.

* Updated option names. Set frequency to drop down. Tweaked functions. Shifted class properties. Introduced set_warning_type().

* Remove ignore line

* added to error message in email

* Remove debug log

* Tweak gitignore

* Clean up error message text

* test

* Revert

* fixed backwards compatiblity issue with arrays

* little modification of string on dashboard

* started  email a cron task

* started  email a cron task

* readded publics

* Revert class-options spacing, cron etc.

* Tweak email cron task.

* Set PHP version to static variable. Extend parent class with subclasses. Fix [] to array() in class-filters.php

* Prevent emails from sending on plugin activation and option update.

* Enhance error message with 'security until' and 'support until' strings.

* Add a default fallback, return,  to the warning_type switch

* made alert dismissable

* Remove hardcoded php version

* Major changes (#11)

* add php version to admin footer. Update grunt file. Update readme.md

* Update readme.md

* email function

* email function

* re-converted option #1 into a checkbox

* Convert php_notifier_send_email option to boolean

* Return early, prevent email error when up to date version of PHP. Resolves #6. Update readme.md

* Add grunt menu, update grunt tasks, test grunt tasks.

* Updated option names. Set frequency to drop down. Tweaked functions. Shifted class properties. Introduced set_warning_type().

* Remove ignore line

* added to error message in email

* Remove debug log

* Tweak gitignore

* Clean up error message text

* test

* Revert

* fixed backwards compatiblity issue with arrays

* little modification of string on dashboard

* started  email a cron task

* started  email a cron task

* readded publics

* Revert class-options spacing, cron etc.

* Tweak email cron task.

* Set PHP version to static variable. Extend parent class with subclasses. Fix [] to array() in class-filters.php

* Prevent emails from sending on plugin activation and option update.

* Enhance error message with 'security until' and 'support until' strings.

* Add a default fallback, return,  to the warning_type switch

* made alert dismissable

* Remove hardcoded php version

+++ php-notifier.php
@@ -17,31 +17,37 @@ class CP_PHP_Notifier {
 	 *
 	 * @var integer
 	 */
-	private $php_version;
+	protected static $php_version;
 
 	/**
 	 * PHP Support Data
 	 *
 	 * @var array
 	 */
-	private $php_support_data;
+	protected $php_support_data;
 
 	/**
-	 * Warning type
+	 * PHP Notifier Settings Array
 	 *
-	 * @var string
+	 * @var array
 	 */
-	private $warning_type;
+	protected static $options;
 
 	public function __construct() {
 
-		define( 'PHP_NOTIFIER_PATH', plugin_dir_path( __FILE__ ) );
-		define( 'PHP_NOTIFIER_URL', plugin_dir_url( __FILE__ ) );
+		define( 'PHP_NOTIFIER_PATH',    plugin_dir_path( __FILE__ ) );
+		define( 'PHP_NOTIFIER_URL',     plugin_dir_url( __FILE__ ) );
 		define( 'PHP_NOTIFIER_VERSION', '1.0.0' );
 
-		$this->php_version = phpversion();
+		self::$php_version = phpversion();
 
 		$this->php_support_data = $this->php_notifier_version_info();
+
+		self::$options = get_option( 'php_notifier_settings', [
+			'send_email'      => true,
+			'email_frequency' => 'monthly',
+			'warning_type'    => false,
+		] );
 
 		$this->init();
 
@@ -59,6 +65,12 @@ class CP_PHP_Notifier {
 		include_once( plugin_dir_path( __FILE__ ) . '/library/partials/class-options.php' );
 
 		include_once( plugin_dir_path( __FILE__ ) . '/library/partials/class-filters.php' );
+
+		include_once( plugin_dir_path( __FILE__ ) . '/library/partials/class-email-cron.php' );
+
+		register_activation_hook( __FILE__, array( $this, 'plugin_activation' ) );
+
+		register_deactivation_hook( __FILE__, array( $this, 'plugin_deactivation' ) );
 
 	}
 
@@ -79,13 +91,13 @@ class CP_PHP_Notifier {
 		//  - Less than last PHP version supported
 		//  - Current time is greater than or equal to "Security Support Until"
 		if (
-			version_compare( $this->php_version, key( $this->php_support_data ), '<' ) ||
+			version_compare( self::$php_version, key( $this->php_support_data ), '<' ) ||
 			strtotime( 'now' ) >= $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ]['security_until']
 		) {
 
-			$this->warning_type = 'deprecated';
+			$this->set_warning_type( 'deprecated' );
 
-			add_action( 'admin_notices', [ $this, 'php_version_error' ] );
+			$this->php_version_error();
 
 			return;
 
@@ -101,24 +113,48 @@ class CP_PHP_Notifier {
 				strtotime( 'now' ) <= $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ]['security_until']
 			) {
 
-				$this->warning_type = 'unsupported';
+				$this->set_warning_type( 'unsupported' );
 
-				add_action( 'admin_notices', [ $this, 'php_version_error' ] );
+				$this->php_version_error();
 
 				return;
 
-			} // @codingStandardsIgnoreLine
+			}
 
 			// PHP Version will not actively be supported in 1 month or less
 			if ( strtotime( '+1 month' ) >= $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ]['supported_until'] ) {
 
-				$this->warning_type = 'deprecated-soon';
+				$this->set_warning_type( 'deprecated-soon' );
 
-				add_action( 'admin_notices', [ $this, 'php_version_error' ] );
+				$this->php_version_error();
 
-			} // @codingStandardsIgnoreLine
+			}
+
+			$this->set_warning_type( false );
 
 		}
+
+	}
+
+	/**
+	 * Set the warning type
+	 *
+	 * @param string $type The type of warning
+	 *        possible: deprecated, unsupported, deprecated-soon
+	 *
+	 * @since 1.0.0
+	 */
+	public function set_warning_type( $type = false ) {
+
+		if ( ! isset( self::$options['warning_type'] ) ) {
+
+			return;
+
+		}
+
+		self::$options['warning_type'] = $type;
+
+		update_option( 'php_notifier_settings', self::$options );
 
 	}
 
@@ -127,16 +163,52 @@ class CP_PHP_Notifier {
 	 *
 	 * @return mixed
 	 */
-	public function php_version_error() {
+	public function php_version_error( $echo = true ) {
 
-		$type = 'error';
+		if ( ! self::$options['warning_type'] ) {
 
-		switch ( $this->warning_type ) {
+			return;
 
-			default:
+		}
+
+		$type            = 'error';
+		$supported_until = isset( $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ] ) ? $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ]['supported_until'] : false;
+		$security_until  = isset( $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ] ) ? $this->php_support_data[ PHP_MAJOR_VERSION . '.' . PHP_MINOR_VERSION ]['security_until'] : false;
+		$additional      = '';
+
+		if ( $supported_until ) {
+
+			$additional .= ( strtotime( 'now' ) > $supported_until ) ? sprintf(
+				'<p>' . __( 'PHP %1$s was officially no longer supported on %2$s.', 'php-notifier' ) . '</p>',
+				esc_html( 'v' . self::$php_version ),
+				$supported_until
+			) : "\r\n\r\n" . sprintf(
+				'<p>' . __( 'PHP %1$s will no longer be supported on %2$s.', 'php-notifier' ) . '</p>',
+				esc_html( 'v' . self::$php_version ),
+				date( get_option( 'date_format' ), $supported_until )
+			);
+
+		}
+
+		if ( $security_until ) {
+
+			$additional .= ( strtotime( 'now' ) > $security_until ) ? sprintf(
+				'<p>' . __( 'PHP %1$s stopped receiving security updates on %2$s.', 'php-notifier' ) . '</p>',
+				esc_html( 'v' . self::$php_version ),
+				$security_until
+			) : "\r\n\r\n" . sprintf(
+				'<p>' . __( 'PHP %1$s will no longer receive security updates on %2$s.', 'php-notifier' ) . '</p>',
+				esc_html( 'v' . self::$php_version ),
+				date( get_option( 'date_format' ), $security_until )
+			);
+
+		}
+
+		switch ( self::$options['warning_type'] ) {
+
 			case 'deprecated':
 
-				$message = __( 'You are running PHP %s, which is deprecated and no longer supported. This is a major security issue and should be addressed immediately. It is highly recommended that you update the version of PHP on your hosting account.', 'php-notifier' );
+				$message    = __( 'You are running PHP %s, which is deprecated and no longer supported. This is a major security issue and should be addressed immediately. It is highly recommended that you update the version of PHP on your hosting account.', 'php-notifier' );
 
 				break;
 
@@ -154,18 +226,36 @@ class CP_PHP_Notifier {
 
 				break;
 
+			default:
+
+				return;
+
+				break;
+
 		}
 
-		printf(
+		$notice = sprintf(
 			'<div class="notice notice-%1$s">
 				<p>%2$s</p>
+				%3$s
 			</div>',
 			$type,
 			sprintf(
 				$message,
-				wp_kses_post( '<strong>v' . $this->php_version . '</strong>' )
-			)
+				wp_kses_post( '<strong>v' . self::$php_version . '</strong>' )
+			),
+			wp_kses_post( $additional )
 		);
+
+		if ( $echo ) {
+
+			echo $notice;
+
+			return;
+
+		}
+
+		return $notice;
 
 	}
 
@@ -176,7 +266,11 @@ class CP_PHP_Notifier {
 	 */
 	public function php_notifier_version_info() {
 
-		delete_transient( 'php_notifier_verison_info' );
+		if ( WP_DEBUG ) {
+
+			delete_transient( 'php_notifier_verison_info' );
+
+		}
 
 		if ( WP_DEBUG || false === ( $php_version_info = get_transient( 'php_notifier_verison_info' ) ) ) {
 
@@ -188,7 +282,7 @@ class CP_PHP_Notifier {
 
 			}
 
-			$body = wp_remote_retrieve_body( $contents );
+			$body = str_replace( '<link rel="shortcut icon" href="http://php.net/favicon.ico">', '', wp_remote_retrieve_body( $contents ) );
 
 			$dom = new DOMDocument;
 
@@ -242,6 +336,36 @@ class CP_PHP_Notifier {
 		}
 
 		return $php_version_info;
+
+	}
+
+	/**
+	 * Plugin Activation
+	 *
+	 * @since 1.0.0
+	 */
+	public function plugin_activation() {
+
+		if ( wp_next_scheduled( 'cp_php_notifier_email' ) || ! self::$options['email_frequency'] ) {
+
+			return;
+
+		}
+
+		update_option( 'php_notifier_prevent_cron', true );
+
+		wp_schedule_event( time(), self::$options['email_frequency'], 'php_notifier_email_cron' );
+
+	}
+
+	/**
+	 * Plugin Deactivation
+	 *
+	 * @since 1.0.0
+	 */
+	public function plugin_deactivation() {
+
+		wp_clear_scheduled_hook( 'php_notifier_email_cron' );
 
 	}
